requirements.txt
pytest==5.4.3
psycopg2-binary==2.8.6
minio==7.0.3
